Fixed: Router Bugs
